+++ app/sections/featured-products/index.tsx
@@ -1,7 +1,4 @@
-import type {
-  ComponentLoaderArgs,
-  HydrogenComponentSchema,
-} from "@weaverse/hydrogen";
+import { type ComponentLoaderArgs, createSchema } from "@weaverse/hydrogen";
 import { forwardRef } from "react";
 import type { FeaturedProductsQuery } from "storefront-api.generated";
 import type { SectionProps } from "~/components/section";
@@ -56,7 +53,7 @@ export let loader = async ({ weaverse }: ComponentLoaderArgs) => {
   );
 };
 
-export let schema: HydrogenComponentSchema = {
+export let schema = createSchema({
   type: "featured-products",
   title: "Featured products",
   childTypes: ["featured-products-items", "heading", "subheading", "paragraph"],
@@ -73,4 +70,4 @@ export let schema: HydrogenComponentSchema = {
       { type: "featured-products-items" },
     ],
   },
-};
+});