+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # @weaverse/pilot
 
+## 5.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.2
+
+## 5.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.1
+
+## 5.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @weaverse/hydrogen@5.2.0
+  - @weaverse/schema@0.4.0
+
 ## 5.0.0
 
 ### Patch Changes